+++ packages/form/src/utils/getElementTypeSymbol.js
@@ -1,0 +1,13 @@
+import { isValidElement } from 'react';
+
+function getElementTypeSymbol(element) {
+    let symbol;
+
+    if (isValidElement(element)) {
+        symbol = element.type.typeSymbol;
+    }
+
+    return symbol || null;
+}
+
+export default getElementTypeSymbol;